pyproject.toml
[tool.poetry]
name = "tkmedia"
version = "0.0.1"
authors = [
    "Marusoftware <marusoftware@outlook.jp>"
]
description = "tkinter media support."

packages = [
    { include="media", from="."}
]

[tool.poetry.dependencies]
python = "^3.7"
av = "^8.0.3"
Pillow = "^8.1.0"
sounddevice = "^0.4.1"
numpy = "^1.20.1"

[tool.poetry.dev-dependencies]
pylint = "^2.6.0"

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"